Work around a Poetry bug
Poetry does not include the src directory in the sdist archive, if a parent
directory of the project also contains a Poetry project. This is the case
because pytest-datadir copies the example project to a temporary directory,
and nox_poetry sets TMPDIR to a location under the session directory, via
session.create_tmp().