Add test for non-ASCII characters in pyproject.toml (#234)
poetry.Config incorrectly used Path.read_text without specifying UTF-8 as the
encoding. On systems with a non-UTF-8 locale, this can lead to decoding errors
because the preferred encoding on the system would be used to decode the
contents of pyproject.toml. This will result in garbled configuration values or
a UnicodeDecodeError exception. For example, the character `Á` will fail to
decode on Windows systems with CP-1252 as the preferred locale encoding.